@@ -1,31 +1,23 @@
-import React, { Component } from 'react';
-import { Todo } from './todo.model';
-import './TodoList.css';
-import { TodoItem } from './TodoItem';
-import { TodoListener, FilterType } from './TodoApp';
-
-// export type TodoListener =  (todo: Todo) => void;
+import { useMemo } from "react";
+import { Todo, TodoStatus } from "./todo.model";
+import { FilterType, TodoListener } from "./TodoApp";
+import TodoItem from "./TodoItem";
+import './TodoList.css'
 
 interface Props {
-  todos: Todo[];
-  filter: FilterType;
-  onChangeStatus: TodoListener;
-  onUpdate: TodoListener;
-  onDelete: TodoListener;
+    todos: Todo[];
+    filter: FilterType;
+    onUpdate: TodoListener;
+    onDelete: TodoListener;
 }
-interface State {}
 
-export default class TodoList extends Component<Props, State> {
-  state = {};
-
-  render() {
-    const {todos, filter, ...rest} = this.props;
-    return (
-      <div className="TodoList">
-        {todos.filter(todo => !filter ? true: todo.status === filter).map((todo) => (
-          <TodoItem key={todo.id} todo={todo} {...rest} />
-        ))}
-      </div>
-    );
-  }
+export default function TodoList({ todos, filter, ...rest }: Props) {
+    const visibleTodos = (todos: Todo[], filter: FilterType) => todos.filter(todo => !filter ? true : todo.status === filter);
+    const memizedVisibleTodos = useMemo(() => visibleTodos(todos, filter), [todos, filter]);
+    return (<div className="TodoList">
+        {
+            memizedVisibleTodos.map(todo =>
+                <TodoItem todo={todo} key={todo.id} {...rest} />)
+        }
+    </div>)
 }
