@@ -2,12 +2,14 @@ import { Todo } from "./todo.model";
 
 const MOCK_TODOS = [
     new Todo('Create TODO App using create-react-app scripts'),
+    new Todo('Create Todo model'),
     new Todo('Create TodoApp component'),
     new Todo('Create TodoList component'),
     new Todo('Create TodoItem component'),
     new Todo('Create TodoInput component'),
-    new Todo('Debug the Application'),
+    new Todo('Debug the application'),
     new Todo('Add Todo filtering by status'),
     new Todo('Improve styling'),
-]
+];
+
 export default MOCK_TODOS;
